requirements.txt
ultralytics
opencv-python
numpy
scikit-learn
matplotlib
PyYAML
